@@ -1,115 +1,91 @@
-import React, { Component } from "react";
-import PropTypes from "prop-types";
+import React, { useState, useEffect } from "react";
 import axios from "axios";
 
-export class Transaction extends Component {
-  state = {
-    loading: true,
-    loaded: false,
-    transaction: null
+const URL_TRANSACTION_BY_ID = `http://localhost:5000/api/v1/account/transaction/`
+
+const Transaction = ({ id }) => {
+  const [loading, setLoading] = useState(false);
+  const [transaction, setTransaction] = useState({});
+
+  useEffect(() => {
+    fetchTransactionDetailById();
+  }, []);
+
+  const fetchTransactionDetailById = async () => {
+    try {
+      setLoading(true);
+      const res = await axios.get(`${URL_TRANSACTION_BY_ID}${id}`);
+      setTransaction(res.data);
+    } catch (err) {
+      console.log(err);
+      setLoading(false);
+    }
   };
 
-  componentDidMount() {
-    axios
-      .get(`http://localhost:5000/api/v1/account/transaction/${this.props.id}`)
-      .then(response => {
-        if (response.status >= 400) {
-          throw response;
-        }
-
-        this.setState({
-          ...this.state,
-          error: null,
-          transaction: response.data
-        });
-      })
-      .catch(err => {
-        switch (err.status) {
-          case 400:
-            return this.setState({
-              ...this.state,
-              error:
-                "Error: Wrong params, make sure are consulting the transaction correctly"
-            });
-          case 404:
-            return this.setState({
-              ...this.state,
-              error: "Error: Transaction not found"
-            });
-          case 500:
-          default:
-            this.setState({
-              ...this.state,
-              error: "Error: Internal server error"
-            });
-            this.setState({
-              ...this.state,
-              error: "Error: Internal server error"
-            });
-        }
-      })
-      .finally(() =>
-        this.setState({ ...this.state, loading: false, loaded: true })
-      );
-  }
-
-  render() {
-    const { loading, loaded, transaction, error } = this.state;
-
-    if (loading || !loaded) {
-      return null;
-    }
-
-    if (error) {
-      return (
-        <div className="alert alert-danger m-2" role="alert">
-          {error}
-        </div>
-      );
-    }
-
+  const error = null;
+  if (error) {
     return (
-      <form className="p-2">
-        {transaction.title && (
-          <div className="form-group">
-            <label htmlFor="title">Title</label>
-            <input
-              type="email"
-              className="form-control"
-              id="title"
-              value={transaction.title}
-              disabled
-            />
-          </div>
-        )}
-
-        {transaction.description && (
-          <div className="form-group">
-            <label htmlFor="description">description</label>
-            <textarea
-              className="form-control"
-              id="description"
-              rows="3"
-              value={transaction.description}
-              disabled
-            />
-          </div>
-        )}
-
-        <div className="form-group">
-          <label>card type</label>
-          <select className="form-control" value={transaction.type} disabled>
-            <option value="credit">credit</option>
-            <option value="debit">debit</option>
-          </select>
-        </div>
-      </form>
+      <div className="alert alert-danger m-2" role="alert">
+        {error}
+      </div>
     );
   }
-}
 
-Transaction.propTypes = {
-  id: PropTypes.string.isRequired
+  return (
+    <form className="p-2">
+        <div className="form-group">
+          <label htmlFor="title">TxId</label>
+          <input
+            type="email"
+            className="form-control"
+            id="title"
+            value={transaction.id}
+            disabled
+          />
+        </div>
+
+        <div className="form-group">
+          <label htmlFor="description">amount</label>
+          <textarea
+            className="form-control"
+            id="description"
+            rows="3"
+            value={transaction.amount}
+            disabled
+          />
+        </div>
+
+        <div className="form-group">
+          <label htmlFor="description">Effective Date</label>
+          <textarea
+            className="form-control"
+            id="description"
+            rows="3"
+            value={transaction.effectiveDate}
+            disabled
+          />
+        </div>
+
+        <div className="form-group">
+          <label htmlFor="description">Status</label>
+          <textarea
+            className="form-control"
+            id="description"
+            rows="3"
+            value={transaction.status}
+            disabled
+          />
+        </div>
+
+      <div className="form-group">
+        <label>type</label>
+        <select className="form-control" value={transaction.type} disabled>
+          <option value="credit">credit</option>
+          <option value="debit">debit</option>
+        </select>
+      </div>
+    </form>
+  );
 };
 
 export default Transaction;
