@@ -1,65 +1,31 @@
-import React, { createContext, Component } from "react";
-import Axios from "axios";
+import React, { useEffect, useState } from "react";
 
 import Transactions from "./components/Transactions";
 import TransactionForm from "./components/TransactionForm";
+import useTransactions from "./hooks/useTransactions";
 
-export const TransactionContext = createContext();
+const App = () => {
+  // const [transactions, addTransaction] = useTransactions([]);
+  const [transactions, setTransactions] = useState([]);
 
-class App extends Component {
-  state = {
-    data: {
-      transactions: [],
-    },
+  const addTransaction = ({ id, amount, type }) => {
+    setTransactions([...transactions, { id, amount, type }]);
   };
 
-  componentDidMount() {
-    Axios.get("http://localhost:5000/api/v1/account/transactions")
-      .then((response) => {
-        if (response.status >= 400) {
-          throw response;
-        }
-
-        this.setState({
-          ...this.state,
-          data: { ...this.state.data, transactions: response.data },
-        });
-      })
-      .catch((err) => {
-        switch (err.status) {
-          case 400:
-            console.log(err);
-            break;
-          default:
-            console.log(err);
-        }
-      })
-      .finally(() =>
-        this.setState({ ...this.state, loading: false, loaded: true })
-      );
-  }
-
-  handleChange = (data) => {
-    this.setState({ ...this.state, data });
+  const loadTransactions = (transactionList) => {
+    setTransactions([...transactions, ...transactionList]);
   };
 
-  render() {
-    return (
-      <div className="App container">
-        {/* TODO: Implement routing */}
-
-        <TransactionContext.Provider
-          value={{ data: this.state.data, handleChange: this.handleChange }}
-        >
-          <TransactionForm />
-
-          <hr />
-
-          <Transactions />
-        </TransactionContext.Provider>
-      </div>
-    );
-  }
-}
+  return (
+    <div className="container">
+      <TransactionForm addTransaction={addTransaction} />
+      <hr />
+      <Transactions
+        transactions={transactions}
+        loadTransactions={loadTransactions}
+      />
+    </div>
+  );
+};
 
 export default App;
